Move all TCP server related flags within `tcp_server.py` and also move the encryption functionality within TCP base server
