+++ proxy/core/base/tcp_server.py
@@ -12,17 +12,69 @@
 
        tcp
 """
+import ssl
+import socket
 import logging
 import selectors
 
 from abc import abstractmethod
-from typing import Any, Optional
+from typing import Any, Optional, Union
+
+from ...common.flag import flags
+from ...common.utils import wrap_socket
+from ...common.types import Readables, SelectableEvents, Writables
+from ...common.constants import DEFAULT_CERT_FILE, DEFAULT_CLIENT_RECVBUF_SIZE
+from ...common.constants import DEFAULT_KEY_FILE, DEFAULT_SERVER_RECVBUF_SIZE, DEFAULT_TIMEOUT
 
 from ...core.work import Work
 from ...core.connection import TcpClientConnection
-from ...common.types import Readables, SelectableEvents, Writables
 
 logger = logging.getLogger(__name__)
+
+
+flags.add_argument(
+    '--key-file',
+    type=str,
+    default=DEFAULT_KEY_FILE,
+    help='Default: None. Server key file to enable end-to-end TLS encryption with clients. '
+    'If used, must also pass --cert-file.',
+)
+
+flags.add_argument(
+    '--cert-file',
+    type=str,
+    default=DEFAULT_CERT_FILE,
+    help='Default: None. Server certificate to enable end-to-end TLS encryption with clients. '
+    'If used, must also pass --key-file.',
+)
+
+flags.add_argument(
+    '--client-recvbuf-size',
+    type=int,
+    default=DEFAULT_CLIENT_RECVBUF_SIZE,
+    help='Default: ' + str(int(DEFAULT_CLIENT_RECVBUF_SIZE / 1024)) +
+    ' KB. Maximum amount of data received from the '
+    'client in a single recv() operation.',
+)
+
+flags.add_argument(
+    '--server-recvbuf-size',
+    type=int,
+    default=DEFAULT_SERVER_RECVBUF_SIZE,
+    help='Default: ' + str(int(DEFAULT_SERVER_RECVBUF_SIZE / 1024)) +
+    ' KB. Maximum amount of data received from the '
+    'server in a single recv() operation.',
+)
+
+flags.add_argument(
+    '--timeout',
+    type=int,
+    default=DEFAULT_TIMEOUT,
+    help='Default: ' + str(DEFAULT_TIMEOUT) +
+    '.  Number of seconds after which '
+    'an inactive connection must be dropped.  Inactivity is defined by no '
+    'data sent or received by the client.',
+)
 
 
 class BaseTcpServerHandler(Work[TcpClientConnection]):
@@ -55,6 +107,17 @@ class BaseTcpServerHandler(Work[TcpClientConnection]):
             self.work.connection.fileno(),
             self.work.address,
         )
+
+    def initialize(self) -> None:
+        """Optionally upgrades connection to HTTPS,
+        sets ``conn`` in non-blocking mode and initializes
+        HTTP protocol plugins."""
+        conn = self._optionally_wrap_socket(self.work.connection)
+        conn.setblocking(False)
+        # Update client connection reference if connection was wrapped
+        if self._encryption_enabled():
+            self.work = TcpClientConnection(conn=conn, addr=self.work.addr)
+        logger.debug('Handling connection %s' % self.work.address)
 
     @abstractmethod
     def handle_data(self, data: memoryview) -> Optional[bool]:
@@ -139,3 +202,20 @@ class BaseTcpServerHandler(Work[TcpClientConnection]):
                     else:
                         teardown = True
         return teardown
+
+    def _encryption_enabled(self) -> bool:
+        return self.flags.keyfile is not None and \
+            self.flags.certfile is not None
+
+    def _optionally_wrap_socket(
+            self, conn: socket.socket,
+    ) -> Union[ssl.SSLSocket, socket.socket]:
+        """Attempts to wrap accepted client connection using provided certificates.
+
+        Shutdown and closes client connection upon error.
+        """
+        if self._encryption_enabled():
+            assert self.flags.keyfile and self.flags.certfile
+            # TODO(abhinavsingh): Insecure TLS versions must not be accepted by default
+            conn = wrap_socket(conn, self.flags.keyfile, self.flags.certfile)
+        return conn
