+++ proxy/http/handler.py
@@ -16,15 +16,11 @@ import asyncio
 import logging
 import selectors
 
-from typing import Tuple, List, Type, Union, Optional, Any
+from typing import Tuple, List, Type, Optional, Any
 
-from ..common.flag import flags
-from ..common.utils import wrap_socket
 from ..core.base import BaseTcpServerHandler
-from ..core.connection import TcpClientConnection
 from ..common.types import Readables, SelectableEvents, Writables
-from ..common.constants import DEFAULT_CLIENT_RECVBUF_SIZE, DEFAULT_KEY_FILE
-from ..common.constants import DEFAULT_SELECTOR_SELECT_TIMEOUT, DEFAULT_TIMEOUT
+from ..common.constants import DEFAULT_SELECTOR_SELECT_TIMEOUT
 
 from .exception import HttpProtocolException
 from .plugin import HttpProtocolHandlerPlugin
@@ -33,32 +29,6 @@ from .parser import HttpParser, httpParserStates, httpParserTypes
 
 
 logger = logging.getLogger(__name__)
-
-
-flags.add_argument(
-    '--client-recvbuf-size',
-    type=int,
-    default=DEFAULT_CLIENT_RECVBUF_SIZE,
-    help='Default: ' + str(int(DEFAULT_CLIENT_RECVBUF_SIZE / 1024)) +
-    ' KB. Maximum amount of data received from the '
-    'client in a single recv() operation.',
-)
-flags.add_argument(
-    '--key-file',
-    type=str,
-    default=DEFAULT_KEY_FILE,
-    help='Default: None. Server key file to enable end-to-end TLS encryption with clients. '
-    'If used, must also pass --cert-file.',
-)
-flags.add_argument(
-    '--timeout',
-    type=int,
-    default=DEFAULT_TIMEOUT,
-    help='Default: ' + str(DEFAULT_TIMEOUT) +
-    '.  Number of seconds after which '
-    'an inactive connection must be dropped.  Inactivity is defined by no '
-    'data sent or received by the client.',
-)
 
 
 class HttpProtocolHandler(BaseTcpServerHandler):
@@ -84,19 +54,6 @@ class HttpProtocolHandler(BaseTcpServerHandler):
     # initialize, is_inactive, shutdown, get_events, handle_events
     # overrides Work class definitions.
     ##
-
-    def initialize(self) -> None:
-        """Optionally upgrades connection to HTTPS,
-        sets ``conn`` in non-blocking mode and initializes
-        HTTP protocol plugins.
-        """
-        conn = self._optionally_wrap_socket(self.work.connection)
-        conn.setblocking(False)
-        # Update client connection reference if connection was wrapped
-        if self._encryption_enabled():
-            self.work = TcpClientConnection(conn=conn, addr=self.work.addr)
-        # self._initialize_plugins()
-        logger.debug('Handling connection %s' % self.work.address)
 
     def is_inactive(self) -> bool:
         if not self.work.has_buffer() and \
@@ -333,23 +290,6 @@ class HttpProtocolHandler(BaseTcpServerHandler):
         )
         self.work._conn = output
         return False
-
-    def _encryption_enabled(self) -> bool:
-        return self.flags.keyfile is not None and \
-            self.flags.certfile is not None
-
-    def _optionally_wrap_socket(
-            self, conn: socket.socket,
-    ) -> Union[ssl.SSLSocket, socket.socket]:
-        """Attempts to wrap accepted client connection using provided certificates.
-
-        Shutdown and closes client connection upon error.
-        """
-        if self._encryption_enabled():
-            assert self.flags.keyfile and self.flags.certfile
-            # TODO(abhinavsingh): Insecure TLS versions must not be accepted by default
-            conn = wrap_socket(conn, self.flags.keyfile, self.flags.certfile)
-        return conn
 
     def _connection_inactive_for(self) -> float:
         return time.time() - self.last_activity
